Creating example to support bug report
When I create this file 'tmp.md' and type the phrase `~~Strike these words out~~.`, I am expecting to see the words struck through as if they appeared in an s tag, like this: <s>Strike these words out</s>. But in the preview pane they do not appear that way. Let's see if they appear properly when I commit the file.